Update README.md
- Fate/Grand Order
FGO is about discovering that humanity is facing extinction, a new recruit to the Chaldea Security Organization goes back in time in order to save the future.

- Fate/Strange Fake
The plot centers around a Grail War faultily copied from the Third Holy Grail War in Fuyuki. After the end of the third Grail War, an organization from the United States that has magi separate from the London-based Mage Association as members took data from Fuyuki's Grail War and planned their own ritual. After seventy years, they used the city Snowfield as the Sacred Land for their own Grail War. They were unable to successfully copy every aspect of the ritual, which led to it acting only as an imitation that has lost the Saber class and allowed for the summoning of strange Servants due to the definition of "hero" being blurred.